CrossSection/AirspaceXSRenderer: add missing signed cast
Without this, the unsigned integer underflows, which results in
strange vertical black lines with the OpenGL renderer.

Closes https://github.com/XCSoar/XCSoar/issues/579

+++ src/CrossSection/AirspaceXSRenderer.cpp
@@ -198,8 +198,8 @@ AirspaceIntersectionVisitorSlice::Render(const AbstractAirspace &as) const
     const PixelSize name_size = canvas.CalcTextSize(name);
     const int x = name_size.width >= max_width
       ? min_x
-      : (min_x + max_x - name_size.width) / 2;
-    const int y = (rcd.top + rcd.bottom - name_size.height) / 2;
+      : (min_x + max_x - (int)name_size.width) / 2;
+    const int y = (rcd.top + rcd.bottom - (int)name_size.height) / 2;
 
     canvas.DrawClippedText({x, y}, max_x - x, name);
   }
